+++ Cmdlets/Public/Set-TeamViewerGroup.ps1
@@ -1,5 +1,6 @@
 function Set-TeamViewerGroup {
     [CmdletBinding(SupportsShouldProcess = $true, DefaultParameterSetName = 'ByParameters')]
+
     param(
         [Parameter(Mandatory = $true)]
         [securestring]
@@ -7,8 +8,8 @@ function Set-TeamViewerGroup {
 
         [Parameter(Mandatory = $true, ValueFromPipeline = $true)]
         [ValidateScript( { $_ | Resolve-TeamViewerGroupId } )]
-        [Alias("GroupId")]
-        [Alias("Id")]
+        [Alias('GroupId')]
+        [Alias('Id')]
         [object]
         $Group,
 
@@ -18,7 +19,7 @@ function Set-TeamViewerGroup {
 
         [Parameter(ParameterSetName = 'ByParameters')]
         [ValidateScript( { $_ | Resolve-TeamViewerPolicyId } )]
-        [Alias("PolicyId")]
+        [Alias('PolicyId')]
         [object]
         $Policy,
 
@@ -26,17 +27,19 @@ function Set-TeamViewerGroup {
         [hashtable]
         $Property
     )
-    Begin {
+
+    begin {
         # Warning suppresion doesn't seem to work.
         # See https://github.com/PowerShell/PSScriptAnalyzer/issues/1472
         $null = $Property
 
         $body = @{}
+
         switch ($PSCmdlet.ParameterSetName) {
             'ByParameters' {
                 $body['name'] = $Name
                 if ($Policy) {
-                    $body['policy_id'] = $Policy | Resolve-TeamViewerPolicyId
+                    $body['policy_id'] = ($Policy | Resolve-TeamViewerPolicyId).ToString()
                 }
             }
             'ByProperties' {
@@ -48,23 +51,23 @@ function Set-TeamViewerGroup {
 
         if ($body.Count -eq 0) {
             $PSCmdlet.ThrowTerminatingError(
-                ("The given input does not change the group." | `
-                        ConvertTo-ErrorRecord -ErrorCategory InvalidArgument))
+                ('The given input does not change the group.' | `
+                    ConvertTo-ErrorRecord -ErrorCategory InvalidArgument))
         }
     }
-    Process {
+
+    process {
         $groupId = $Group | Resolve-TeamViewerGroupId
         $resourceUri = "$(Get-TeamViewerApiUri)/groups/$groupId"
 
-        if ($PSCmdlet.ShouldProcess($groupId, "Update group")) {
+        if ($PSCmdlet.ShouldProcess($groupId, 'Update group')) {
             Invoke-TeamViewerRestMethod `
                 -ApiToken $ApiToken `
                 -Uri $resourceUri `
                 -Method Put `
-                -ContentType "application/json; charset=utf-8" `
+                -ContentType 'application/json; charset=utf-8' `
                 -Body ([System.Text.Encoding]::UTF8.GetBytes(($body | ConvertTo-Json))) `
-                -WriteErrorTo $PSCmdlet | `
-                Out-Null
+                -WriteErrorTo $PSCmdlet | Out-Null
         }
     }
 }